refactor(models): one export per file in fns/ (LFE-14948)
Split the form-value module into one function per file and move
parsePriceInput out of validation.ts (structure RFC rules 3, 4 and 20);
unexport the schemas nothing outside validation.ts imports.
