+++ web/src/features/models/components/UpsertModelFormDialog/UpsertModelFormDialog.tsx
@@ -33,11 +33,9 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/src/components/ui/select";
-import {
-  buildFormValues,
-  matchPatternFor,
-  toPricingTierInputs,
-} from "@/src/features/models/fns/buildFormValues";
+import { buildFormValues } from "@/src/features/models/fns/buildFormValues";
+import { matchPatternFor } from "@/src/features/models/fns/matchPatternFor";
+import { toPricingTierInputs } from "@/src/features/models/fns/toPricingTierInputs";
 import {
   type FormUpsertModel,
   FormUpsertModelSchema,
