+++ web/src/features/models/fns/buildFormValues.ts
@@ -1,13 +1,11 @@
 import Decimal from "decimal.js";
 
-import { type PricingTierInput } from "@langfuse/shared";
-
+import { matchPatternFor } from "@/src/features/models/fns/matchPatternFor";
 import {
   type FormPricingTier,
   type FormUpsertModel,
   type FormUsageType,
   type GetModelResult,
-  parsePriceInput,
 } from "@/src/features/models/validation";
 
 /** What the dialog knows about the model being created, edited or cloned. */
@@ -26,25 +24,8 @@ const CREATE_DEFAULT_PRICES: Record<string, number> = {
   output: 0.000002,
 };
 
-export const matchPatternFor = (modelName: string) => `(?i)^(${modelName})$`;
-
 /** Full decimal notation — `String(1e-7)` would put "1e-7" in the input. */
 const formatPrice = (price: number) => new Decimal(price).toFixed();
-
-/** Row keys are opaque and only need to be unique within one form instance. */
-export const makeUsageTypeKeys = (
-  existing: Pick<FormUsageType, "key">[],
-  count: number,
-): string[] => {
-  const used = new Set(existing.map((row) => row.key));
-  const keys: string[] = [];
-  let candidate = existing.length;
-  while (keys.length < count) {
-    const key = `u${candidate++}`;
-    if (!used.has(key)) keys.push(key);
-  }
-  return keys;
-};
 
 const toUsageTypeRows = (names: string[]): FormUsageType[] =>
   names.map((name, index) => ({ key: `u${index}`, name }));
@@ -111,32 +92,4 @@ export const buildFormValues = (source: ModelFormSource): FormUpsertModel => {
       prices: pricesByRowKey(usageTypes, tier.prices),
     })),
   };
-};
-
-/** Priority is derived from tier order, so it can never drift or collide. */
-export const derivePriorities = (tiers: { isDefault: boolean }[]): number[] => {
-  let next = 1;
-  return tiers.map((tier) => (tier.isDefault ? 0 : next++));
-};
-
-export const toPricingTierInputs = (
-  values: Pick<FormUpsertModel, "usageTypes" | "pricingTiers">,
-): PricingTierInput[] => {
-  const priorities = derivePriorities(values.pricingTiers);
-
-  return values.pricingTiers.map((tier, tierIndex) => ({
-    name: tier.name,
-    isDefault: tier.isDefault,
-    priority: priorities[tierIndex],
-    conditions: tier.conditions.map((condition) => ({
-      ...condition,
-      caseSensitive: condition.caseSensitive ?? false,
-    })),
-    prices: Object.fromEntries(
-      values.usageTypes.flatMap((row) => {
-        const price = parsePriceInput(tier.prices[row.key]);
-        return price === null ? [] : [[row.name.trim(), price] as const];
-      }),
-    ),
-  }));
 };
