+++ web/src/features/models/fns/matchPatternFor.ts
@@ -1,0 +1,2 @@
+/** The exact, case-insensitive Postgres pattern for one model name. */
+export const matchPatternFor = (modelName: string) => `(?i)^(${modelName})$`;
